+++ investement/src/main/java/com/example/investment_api/virtual/account/controller/AccountController.java
@@ -7,8 +7,10 @@ import com.example.investment_api.virtual.account.controller.dto.*;
 import com.example.investment_api.virtual.account.domain.MemberAccount;
 
 import com.example.investment_api.virtual.account.domain.MemberOrder;
-import com.example.investment_api.virtual.account.service.MemberAccountService;
+import com.example.investment_api.virtual.account.service.AccountService;
+import com.example.investment_api.virtual.account.service.OrderService;
 
+import com.example.investment_api.virtual.account.service.StockDataTransferService;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -24,10 +26,14 @@ import java.util.List;
 @RequestMapping("/api/account")
 public class AccountController {
 
-    private final MemberAccountService memberAccountService;
+    private final OrderService memberAccountService;
+    private final StockDataTransferService transferService;
+    private final AccountService accountService;
 
-    public AccountController(MemberAccountService memberAccountService) {
+    public AccountController(OrderService memberAccountService, StockDataTransferService transferService, AccountService accountService) {
         this.memberAccountService = memberAccountService;
+        this.transferService = transferService;
+        this.accountService = accountService;
     }
 
     @PostMapping("/buy")
@@ -59,18 +65,18 @@ public class AccountController {
 
     @GetMapping("/accounts")
     public List<MemberAccount> getMemberAccounts(@Member Long memberId) {
-        return memberAccountService.getMemberAccounts(memberId);
+        return accountService.getMemberAccounts(memberId);
     }
 
     @GetMapping("/accounts/{stockName}")
     public MemberAccount getMemberAccount(@Member Long memberId,
                                           @PathVariable String stockName) {
-        return memberAccountService.getMemberAccount(memberId, stockName);
+        return accountService.getMemberAccount(memberId, stockName);
     }
 
     @GetMapping("/orders/{stockName}")
-    public List<OrderData> getMemberOrders(@Member Long memberId, @PathVariable String stockName){
-        return memberAccountService.getStockOrderData(memberId, stockName);
+    public List<OrderData> getMemberOrders(@Member Long memberId, @PathVariable String stockName) {
+        return transferService.getStockOrderData(memberId, stockName);
     }
 
 
@@ -87,7 +93,7 @@ public class AccountController {
     }
 
     @GetMapping("accounts/save/{stockName}")
-    public List<MemberOrder> getMemberOrdersList(@Member Long memberId, @PathVariable String stockName){
+    public List<MemberOrder> getMemberOrdersList(@Member Long memberId, @PathVariable String stockName) {
         return memberAccountService.getMemberOrders(memberId, stockName);
     }
 }
