+++ investement/src/main/java/com/example/investment_api/virtual/calculator/controller/StockCalculatorController.java
@@ -4,7 +4,7 @@ import com.example.investment_api.global.annotation.Member;
 import com.example.investment_api.virtual.account.controller.dto.ResultDTO;
 import com.example.investment_api.virtual.account.controller.dto.AccountStockData;
 import com.example.investment_api.virtual.account.controller.dto.UserStockDTO;
-import com.example.investment_api.virtual.account.service.MemberAccountService;
+import com.example.investment_api.virtual.account.service.OrderService;
 import com.example.investment_api.virtual.account.service.StockDataTransferService;
 import com.example.investment_api.virtual.calculator.mapper.StockCalculationMapper;
 import org.springframework.http.HttpStatus;
@@ -21,17 +21,15 @@ public class StockCalculatorController {
 
     private final StockCalculationMapper stockCalculationMapper;
     private final StockDataTransferService stockDataTransferService;
-    private final MemberAccountService memberAccountService;
 
-    public StockCalculatorController(StockCalculationMapper stockCalculationService, StockDataTransferService stockDataTransferService, MemberAccountService memberAccountService) {
+    public StockCalculatorController(StockCalculationMapper stockCalculationService, StockDataTransferService stockDataTransferService) {
         this.stockCalculationMapper = stockCalculationService;
         this.stockDataTransferService = stockDataTransferService;
-        this.memberAccountService = memberAccountService;
     }
 
     @GetMapping()
     public ResponseEntity<List<ResultDTO>> getAllCalculations(@Member Long memberId) {
-        UserStockDTO userStockDTO = memberAccountService.getUserStockNames(memberId);
+        UserStockDTO userStockDTO = stockDataTransferService.getUserStockNames(memberId);
         List<String> stockNames = userStockDTO.stockNames();
 
         List<ResultDTO> results = stockNames.stream()
